backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-generativeai
Pillow
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
alembic
passlib
asyncpg
